Add: project directory restructure and include the spinner component
